feature: set screen presentation between intro and addMember
