--- 6Bubbles/Views/TestView.swift
@@ -1,20 +1,0 @@
-//
-//  TestView.swift
-//  6Bubbles
-//
-//  Created by 김민 on 2023/05/07.
-//
-
-import SwiftUI
-
-struct TestView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct TestView_Previews: PreviewProvider {
-    static var previews: some View {
-        TestView()
-    }
-}
